+++ book_resale_finder/__init__.py
@@ -1,3 +1,3 @@
 """Book Resale Finder."""
 
-__version__ = "1.1.3"
+__version__ = "1.1.4"